(SIMP-MAINT) Fix unecessary package logic
A last-minute change to the `pkg_rpm` PR introduced a debian-specific,
assumed-to-be-root logic that always ran, but would never be appropriate
to run because the job was already tagged with 'rpmbuild', which limits
the job to rpmbuild-capable runners.

This patch adds logic that tests for the existance of the rpmbuild
command before attempting remedial measures, and will fail if the
`rpmbuild` command is unavailable (also missing from the earlier logic).

None of this logic will be necessary unless the `rpmbuild` tag is
removed from the CI job definition.