(SIMP-10575) Make compatible with pkg:single
pkg-r10k generates its .spec file from an .erb file, which causes
`pkg:single` (used by GHA's `relase_rpms` workflow) to fail during its
`rpm_rebuild?` checks.  The logic of `pkg:single` is monstrous and so
are setup of of its spec tests.  So, as a fast workaround that enabled
GHA to build and sign RPMs, this patch simply adds the `.spec` file to
the repo.

During `pkg:rpm`, the tracked `.spec` will be overwritten, based on the
contents of `build/sources.yaml`.

[SIMP-10575] #close